#154 [move] : move set meet up detail color extensions to SetDday.kt
